Add functions
* Added functions to calculate the AveragePrice
* Bound state values on the view
* Added saved newItem on the userDefaults

+++ AveragePriceCalculator/Tools/Extension/Extension+Double.swift
@@ -1,5 +1,5 @@
 //
-//  Extension+Float.swift
+//  Extension+Double.swift
 //  AveragePriceCalculator
 //
 //  Created by Jiwon Yoon on 8/17/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Float {
+extension Double {
 
     func displayDecimalPlace(by: Int) -> String {
         String(format: "%.\(by)f", self)
